2304.04912v1.pdf
Financial Time Series Forecasting using CNN and Transformer
 Zhen Zeng, Rachneet Kaur, Suchetha Siddagangappa,
 Saba Rahimi, Tucker Balch, Manuela Veloso
 J. P. Morgan AI Research, New York, NY, USA

Abstract
Time series forecasting is important across various domains
for decision-making. In particular, financial time series such
as stock prices can be hard to predict as it is difficult to
model short-term and long-term temporal dependencies between data points. Convolutional Neural Networks (CNN) are
good at capturing local patterns for modeling short-term dependencies. However, CNNs cannot learn long-term dependencies due to the limited receptive field. Transformers on the
other hand are capable of learning global context and longterm dependencies. In this paper, we propose to harness the
power of CNNs and Transformers to model both short-term
and long-term dependencies within a time series, and forecast
if the price would go up, down or remain the same (flat) in
the future. In our experiments, we demonstrated the success
of the proposed method in comparison to commonly adopted
statistical and deep learning methods on forecasting intraday
stock price change of S&P 500 constituents.